Add Extensions
